Reorganize CacheHandler files
Instead of separating the scanning and serialization functions across multiple files, place them all in the main CacheHandler file, separating them through `#region`s.
